fix(CallExpr) : now passing the expressions as arguments to the call function instead of LiteralValues
earlier the evaluated LiteralValues were passed as args to the call
method of the callable class which was not sufficient for utilizing the
resolved variables and functions in the locals map. Now passing
expressions we can directly access the resolved variables without
wanting to resolve again for nativeFunctions like typOf function.